+++ desktop/src/main/java/com/onecat/burst/ui/displays/DisplayComponent.java
@@ -3,6 +3,7 @@ package com.onecat.burst.ui.displays;
 import com.badlogic.gdx.graphics.g3d.particles.emitters.RegularEmitter;
 import com.badlogic.gdx.graphics.g3d.particles.values.RangedNumericValue;
 import com.badlogic.gdx.graphics.g3d.particles.values.ScaledNumericValue;
+import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
@@ -14,13 +15,12 @@ import java.util.Locale;
 public abstract class DisplayComponent extends Table {
 
 	public static final float DEFAULT_SPACING = 4f;
-	public static final float NAME_LABEL_WIDTH = 90f;
-	public static final float RANGE_LABEL_WIDTH = 64f;
-	public static final float INPUT_FIELD_WIDTH = 64f;
+	public static final float NAME_LABEL_WIDTH = 80f;
+	public static final float INPUT_FIELD_WIDTH = 52f;
 
 	public DisplayComponent(Skin skin) {
 		super(skin);
-		defaults().space(DEFAULT_SPACING).expandX();
+		applyTableDefaults(this);
 		setBackground(skin.getDrawable("button_flat_default_disabled"));
 	}
 
@@ -34,12 +34,12 @@ public abstract class DisplayComponent extends Table {
 			super(skin);
 			this.emitter = emitter;
 			Label rangeLabel = new Label("Count", skin);
-			add(rangeLabel).width(NAME_LABEL_WIDTH);
+			applyComponentNameSizing(add(rangeLabel));
 			Label minLabel = new Label("min", skin.get("small", Label.LabelStyle.class));
 			minLabel.setAlignment(Align.center);
-			add(minLabel).width(RANGE_LABEL_WIDTH);
+			add(minLabel);
 			oldMin = String.valueOf(emitter.minParticleCount);
-			TextField minField = new TextField(oldMin, skin);
+			TextField minField = textField(oldMin, skin);
 			minField.setTextFieldFilter(UserInput.integerFilter());
 			minField.addListener(UserInput.confirmInputListener(minField));
 			minField.addListener(new FocusListener() {
@@ -50,12 +50,12 @@ public abstract class DisplayComponent extends Table {
 						onMinCountEdited(minField);
 				}
 			});
-			add(minField).width(INPUT_FIELD_WIDTH);
+			add(minField);
 			Label maxLabel = new Label("max", skin.get("small", Label.LabelStyle.class));
 			maxLabel.setAlignment(Align.center);
-			add(maxLabel).width(RANGE_LABEL_WIDTH);
+			add(maxLabel);
 			oldMax = String.valueOf(emitter.maxParticleCount);
-			TextField maxField = new TextField(oldMax, skin);
+			TextField maxField = textField(oldMax, skin);
 			maxField.setTextFieldFilter(UserInput.integerFilter());
 			maxField.addListener(UserInput.confirmInputListener(maxField));
 			maxField.addListener(new FocusListener() {
@@ -66,7 +66,7 @@ public abstract class DisplayComponent extends Table {
 						onMaxCountEdited(maxField);
 				}
 			});
-			add(maxField).width(INPUT_FIELD_WIDTH);
+			add(maxField);
 		}
 
 		private void onMinCountEdited(TextField inputField) {
@@ -104,17 +104,23 @@ public abstract class DisplayComponent extends Table {
 		private String oldMin;
 		private String oldMax;
 		private final RangedNumericValue value;
+		private final boolean forcePositive;
+		private final boolean forceInteger;
+		private final boolean autoActivate;
 
-		public RangedNumericComponent(String name, RangedNumericValue value, Skin skin) {
+		public RangedNumericComponent(String name, RangedNumericValue value, Skin skin, boolean forcePositive, boolean forceInteger, boolean autoActivate) {
 			super(skin);
 			this.value = value;
+			this.forcePositive = forcePositive;
+			this.forceInteger = forceInteger;
+			this.autoActivate = autoActivate;
 			Label rangeLabel = new Label(String.format(Locale.ROOT, "%s:", name), skin);
-			add(rangeLabel).width(NAME_LABEL_WIDTH);
+			applyComponentNameSizing(add(rangeLabel));
 			Label minLabel = new Label("min", skin.get("small", Label.LabelStyle.class));
 			minLabel.setAlignment(Align.center);
-			add(minLabel).width(RANGE_LABEL_WIDTH);
-			oldMin = String.valueOf(value.getLowMin());
-			TextField minField = new TextField(oldMin, skin);
+			add(minLabel);
+			oldMin = forceInteger ? String.valueOf(MathUtils.round(value.getLowMin())) : String.valueOf(value.getLowMin());
+			TextField minField = textField(oldMin, skin);
 			minField.setTextFieldFilter(UserInput.floatFilter());
 			minField.addListener(UserInput.confirmInputListener(minField));
 			minField.addListener(new FocusListener() {
@@ -125,12 +131,12 @@ public abstract class DisplayComponent extends Table {
 						onMinCountEdited(minField);
 				}
 			});
-			add(minField).width(INPUT_FIELD_WIDTH);
+			add(minField);
 			Label maxLabel = new Label("max", skin.get("small", Label.LabelStyle.class));
 			maxLabel.setAlignment(Align.center);
-			add(maxLabel).width(RANGE_LABEL_WIDTH);
-			oldMax = String.valueOf(value.getLowMax());
-			TextField maxField = new TextField(oldMax, skin);
+			add(maxLabel);
+			oldMax = forceInteger ? String.valueOf(MathUtils.round(value.getLowMax())) : String.valueOf(value.getLowMax());
+			TextField maxField = textField(oldMax, skin);
 			maxField.setTextFieldFilter(UserInput.floatFilter());
 			maxField.addListener(UserInput.confirmInputListener(maxField));
 			maxField.addListener(new FocusListener() {
@@ -141,34 +147,48 @@ public abstract class DisplayComponent extends Table {
 						onMaxCountEdited(maxField);
 				}
 			});
-			add(maxField).width(INPUT_FIELD_WIDTH);
+			add(maxField);
+			if (!autoActivate) {
+				CheckBox activeCheckbox = new CheckBox("Active", skin);
+				activeCheckbox.setChecked(value.isActive());
+				activeCheckbox.addListener(new ChangeListener() {
+
+					@Override
+					public void changed(ChangeEvent event, Actor actor) {
+						value.setActive(activeCheckbox.isChecked());
+					}
+				});
+				add(activeCheckbox);
+			}
 		}
 
 		private void onMinCountEdited(TextField inputField) {
-			float newValue;
 			if (UserInput.isFloat(inputField.getText())) {
-				newValue = Float.parseFloat(inputField.getText());
-				if (newValue >= 0) {
-					value.setLowMin(newValue);
-					oldMin = String.valueOf(newValue);
+				float newValue = Float.parseFloat(inputField.getText());
+				boolean isValid = !forcePositive || !(newValue < 0);
+				if (isValid) {
+					value.setLowMin(forceInteger ? MathUtils.round(newValue) : newValue);
+					oldMin = forceInteger ? String.valueOf(MathUtils.round(newValue)) : String.valueOf(newValue);
+					inputField.setText(oldMin);
+					if (autoActivate) value.setActive(true);
 					return;
 				}
 			}
-			// Invalid input
 			inputField.setText(oldMin);
 		}
 
 		private void onMaxCountEdited(TextField inputField) {
-			float newValue;
 			if (UserInput.isFloat(inputField.getText())) {
-				newValue = Float.parseFloat(inputField.getText());
-				if (newValue >= 0) {
-					value.setLowMax(newValue);
-					oldMax = String.valueOf(newValue);
+				float newValue = Float.parseFloat(inputField.getText());
+				boolean isValid = !forcePositive || !(newValue < 0);
+				if (isValid) {
+					value.setLowMax(forceInteger ? MathUtils.round(newValue) : newValue);
+					oldMax = forceInteger ? String.valueOf(MathUtils.round(newValue)) : String.valueOf(newValue);
+					inputField.setText(oldMax);
+					if (autoActivate) value.setActive(true);
 					return;
 				}
 			}
-			// Invalid input
 			inputField.setText(oldMax);
 		}
 
@@ -181,19 +201,25 @@ public abstract class DisplayComponent extends Table {
 		private String oldLowMin;
 		private String oldLowMax;
 		private final ScaledNumericValue value;
+		private final boolean forcePositive;
+		private final boolean forceInteger;
+		private final boolean autoActivate;
 
-		public ScaledNumericComponent(String name, ScaledNumericValue value, Skin skin) {
+		public ScaledNumericComponent(String name, ScaledNumericValue value, Skin skin, boolean forcePositive, boolean forceInteger, boolean autoActivate) {
 			super(skin);
 			this.value = value;
+			this.forcePositive = forcePositive;
+			this.forceInteger = forceInteger;
+			this.autoActivate = autoActivate;
 			Label rangeLabel = new Label(String.format(Locale.ROOT, "%s:", name), skin);
-			add(rangeLabel).width(NAME_LABEL_WIDTH);
+			DisplayComponent.applyComponentNameSizing(add(rangeLabel));
 			Table nestedTable = new Table();
-			nestedTable.defaults().space(DEFAULT_SPACING).expandX();
+			applyTableDefaults(nestedTable);
 			Label highMinLabel = new Label("high min", skin.get("small", Label.LabelStyle.class));
 			highMinLabel.setAlignment(Align.center);
-			nestedTable.add(highMinLabel).width(RANGE_LABEL_WIDTH);
-			oldHighMin = String.valueOf(value.getHighMin());
-			TextField highMinField = new TextField(oldHighMin, skin);
+			nestedTable.add(highMinLabel);
+			oldHighMin = forceInteger ? String.valueOf(MathUtils.round(value.getHighMin())) : String.valueOf(value.getHighMin());
+			TextField highMinField = textField(oldHighMin, skin);
 			highMinField.setTextFieldFilter(UserInput.floatFilter());
 			highMinField.addListener(UserInput.confirmInputListener(highMinField));
 			highMinField.addListener(new FocusListener() {
@@ -204,12 +230,12 @@ public abstract class DisplayComponent extends Table {
 						onHighMinEdited(highMinField);
 				}
 			});
-			nestedTable.add(highMinField).width(INPUT_FIELD_WIDTH);
+			nestedTable.add(highMinField);
 			Label highMaxLabel = new Label("high max", skin.get("small", Label.LabelStyle.class));
 			highMaxLabel.setAlignment(Align.center);
-			nestedTable.add(highMaxLabel).width(RANGE_LABEL_WIDTH);
-			oldHighMax = String.valueOf(value.getHighMax());
-			TextField highMaxField = new TextField(oldHighMax, skin);
+			nestedTable.add(highMaxLabel);
+			oldHighMax = forceInteger ? String.valueOf(MathUtils.round(value.getHighMax())) : String.valueOf(value.getHighMax());
+			TextField highMaxField = textField(oldHighMax, skin);
 			highMaxField.setTextFieldFilter(UserInput.floatFilter());
 			highMaxField.addListener(UserInput.confirmInputListener(highMaxField));
 			highMaxField.addListener(new FocusListener() {
@@ -220,13 +246,13 @@ public abstract class DisplayComponent extends Table {
 						onHighMaxEdited(highMaxField);
 				}
 			});
-			nestedTable.add(highMaxField).width(INPUT_FIELD_WIDTH).row();
+			nestedTable.add(highMaxField).row();
 
 			Label lowMinLabel = new Label("low min", skin.get("small", Label.LabelStyle.class));
 			lowMinLabel.setAlignment(Align.center);
-			nestedTable.add(lowMinLabel).width(RANGE_LABEL_WIDTH);
-			oldLowMin = String.valueOf(value.getLowMin());
-			TextField lowMinField = new TextField(oldLowMin, skin);
+			nestedTable.add(lowMinLabel);
+			oldLowMin = forceInteger ? String.valueOf(MathUtils.round(value.getLowMin())) : String.valueOf(value.getLowMin());
+			TextField lowMinField = textField(oldLowMin, skin);
 			lowMinField.setTextFieldFilter(UserInput.floatFilter());
 			lowMinField.addListener(UserInput.confirmInputListener(lowMinField));
 			lowMinField.addListener(new FocusListener() {
@@ -237,12 +263,12 @@ public abstract class DisplayComponent extends Table {
 						onLowMinEdited(lowMinField);
 				}
 			});
-			nestedTable.add(lowMinField).width(INPUT_FIELD_WIDTH);
+			nestedTable.add(lowMinField);
 			Label lowMaxLabel = new Label("low max", skin.get("small", Label.LabelStyle.class));
 			lowMaxLabel.setAlignment(Align.center);
-			nestedTable.add(lowMaxLabel).width(RANGE_LABEL_WIDTH);
-			oldLowMax = String.valueOf(value.getLowMax());
-			TextField lowMaxField = new TextField(oldLowMax, skin);
+			nestedTable.add(lowMaxLabel);
+			oldLowMax = forceInteger ? String.valueOf(MathUtils.round(value.getLowMax())) : String.valueOf(value.getLowMax());
+			TextField lowMaxField = textField(oldLowMax, skin);
 			lowMaxField.setTextFieldFilter(UserInput.floatFilter());
 			lowMaxField.addListener(UserInput.confirmInputListener(lowMaxField));
 			lowMaxField.addListener(new FocusListener() {
@@ -253,8 +279,8 @@ public abstract class DisplayComponent extends Table {
 						onLowMaxEdited(lowMaxField);
 				}
 			});
-			nestedTable.add(lowMaxField).width(INPUT_FIELD_WIDTH);
-			add(nestedTable).row();
+			nestedTable.add(lowMaxField);
+			add(nestedTable).growX().row();
 
 			EditableGraph graph = new EditableGraph(value, skin);
 			add(graph).colspan(2).growX().row();
@@ -272,61 +298,84 @@ public abstract class DisplayComponent extends Table {
 		}
 
 		private void onHighMinEdited(TextField inputField) {
-			float newValue;
 			if (UserInput.isFloat(inputField.getText())) {
-				newValue = Float.parseFloat(inputField.getText());
-				if (newValue >= 0) {
-					value.setHighMin(newValue);
-					oldHighMin = String.valueOf(newValue);
+				float newValue = Float.parseFloat(inputField.getText());
+				boolean isValid = !forcePositive || !(newValue < 0);
+				if (isValid) {
+					value.setHighMin(forceInteger ? MathUtils.round(newValue) : newValue);
+					oldHighMin = forceInteger ? String.valueOf(MathUtils.round(newValue)) : String.valueOf(newValue);
+					inputField.setText(oldHighMin);
+					if (autoActivate) value.setActive(true);
 					return;
 				}
 			}
-			// Invalid input
 			inputField.setText(oldHighMin);
 		}
 
 		private void onHighMaxEdited(TextField inputField) {
-			float newValue;
 			if (UserInput.isFloat(inputField.getText())) {
-				newValue = Float.parseFloat(inputField.getText());
-				if (newValue >= 0) {
-					value.setHighMax(newValue);
-					oldHighMax = String.valueOf(newValue);
+				float newValue = Float.parseFloat(inputField.getText());
+				boolean isValid = !forcePositive || !(newValue < 0);
+				if (isValid) {
+					value.setHighMax(forceInteger ? MathUtils.round(newValue) : newValue);
+					oldHighMax = forceInteger ? String.valueOf(MathUtils.round(newValue)) : String.valueOf(newValue);
+					inputField.setText(oldHighMax);
+					if (autoActivate) value.setActive(true);
 					return;
 				}
 			}
-			// Invalid input
 			inputField.setText(oldHighMax);
 		}
 
 		private void onLowMinEdited(TextField inputField) {
-			float newValue;
 			if (UserInput.isFloat(inputField.getText())) {
-				newValue = Float.parseFloat(inputField.getText());
-				if (newValue >= 0) {
-					value.setLowMin(newValue);
-					oldLowMin = String.valueOf(newValue);
+				float newValue = Float.parseFloat(inputField.getText());
+				boolean isValid = !forcePositive || !(newValue < 0);
+				if (isValid) {
+					value.setLowMin(forceInteger ? MathUtils.round(newValue) : newValue);
+					oldLowMin = forceInteger ? String.valueOf(MathUtils.round(newValue)) : String.valueOf(newValue);
+					inputField.setText(oldLowMin);
+					if (autoActivate) value.setActive(true);
 					return;
 				}
 			}
-			// Invalid input
 			inputField.setText(oldLowMin);
 		}
 
 		private void onLowMaxEdited(TextField inputField) {
-			float newValue;
 			if (UserInput.isFloat(inputField.getText())) {
-				newValue = Float.parseFloat(inputField.getText());
-				if (newValue >= 0) {
-					value.setLowMax(newValue);
-					oldLowMax = String.valueOf(newValue);
+				float newValue = Float.parseFloat(inputField.getText());
+				boolean isValid = !forcePositive || !(newValue < 0);
+				if (isValid) {
+					value.setLowMax(forceInteger ? MathUtils.round(newValue) : newValue);
+					oldLowMax = forceInteger ? String.valueOf(MathUtils.round(newValue)) : String.valueOf(newValue);
+					inputField.setText(oldLowMax);
+					if (autoActivate) value.setActive(true);
 					return;
 				}
 			}
-			// Invalid input
 			inputField.setText(oldLowMax);
 		}
 
+	}
+
+	private static TextField textField(String text, Skin skin) {
+		return new TextField(text, skin) {
+
+			@Override
+			public float getPrefWidth() {
+				// libGDX forces hardcoded 150px by default which fucks up Table layout a bit
+				return INPUT_FIELD_WIDTH;
+			}
+		};
+	}
+
+	private static void applyTableDefaults(Table table) {
+		table.defaults().space(DEFAULT_SPACING).growX();
+	}
+
+	private static void applyComponentNameSizing(Cell<Label> cell) {
+		cell.expand(false, false).width(NAME_LABEL_WIDTH);
 	}
 
 }